[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 
-void sum(int a, int b)
+void sum(long long int a, long long int b)
 {
-    printf("%d",(a+b)*(a-b));
+    printf("%lld", (a + b) * (a - b));
 }
 int main()
 {
-    int a=0, b=0;
-    scanf("%d %d",&a,&b);
-    
-    sum(a,b);
-    
+    long long int a = 0, b = 0;
+    scanf("%lld %lld", &a, &b);
+
+    sum(a, b);
+
     return 0;
-    
+
 }
